+ added localization of strings via Resource Bundles - English and French + moved source file to own directory + setup as an IntelliJ and NetBeans project
