+++ .gitignore
@@ -1,1 +1,3 @@
 /**/*.class
+out/
+/build/
