+++ README.md
@@ -1,5 +1,19 @@
-# Singleton Pattern Program
-# Gordon Doskas
-# PRG/421
-# January 25, 2016
-# Roland Morales
+Singleton Pattern Program - Week 2 Team Improvements
+
+Team B -
+ ( Elijah Cornell / Eric Landeis / Gordon Doskas / James Rippon /
+   Joseph Hart / Keith Green / Lance Branford )
+
+PRG/421
+February 1, 2016
+Roland Morales
+
+-
+
+The project is based on  Gordon's Week 1 program
+
+Team Improvements:
+ -  Refactored program into separate classes
+ -  Added localization of strings via Resource Bundles - English and French
+ -
+ -
