+++ src/Strings_en.properties
@@ -1,0 +1,12 @@
+# Main Menu
+which.lane=Which lane?
+menu.display=[D]isplay Racetrack
+menu.add=[A]dd Runner
+menu.exit=e[X]it
+menu.display.code=D
+menu.add.code=A
+menu.exit.code=X
+#
+# Error Messages
+error.invalid.lane=Invalid lane
+error.lane.occupied=Lane occupied
