+++ src/Strings_fr.properties
@@ -1,0 +1,12 @@
+# Main Menu
+which.lane=Quelle voie?
+menu.display=Affichage [H]ippodrome
+menu.add=[A]jouter Coureur
+menu.exit=[S]ortie
+menu.display.code=H
+menu.add.code=A
+menu.exit.code=S
+#
+# Error Messages
+error.invalid.lane=Voies non valide
+error.lane.occupied=Voie occupée
